+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Controls/AcquiredImages/AcquiredImagesManagerViewVM.cs
@@ -262,6 +262,7 @@ namespace Assistant.NINAPlugin.Controls.AcquiredImages {
         public string ProjectName { get; private set; }
         public string TargetName { get; private set; }
         public bool Accepted { get { return acquiredImage.Accepted; } }
+        public string RejectReason { get { return acquiredImage.RejectReason; } }
 
         public string FileName { get { return acquiredImage.Metadata.FileName; } }
         public string ExposureDuration { get { return fmt(acquiredImage.Metadata.ExposureDuration); } }